general refactoring
use common functions (removeQuote, genFuncArgs),
move/split `util` module,
