+++ lib/freecad_stub_gen/generators/common/return_type_converter/full.py
@@ -4,7 +4,7 @@ from inspect import Signature
 
 from freecad_stub_gen.generators.common.annotation_parameter import RawRepr
 from freecad_stub_gen.generators.common.cpp_function import generateExpressionUntilChar
-from freecad_stub_gen.generators.common.names import validatePythonValue
+from freecad_stub_gen.cpp_code.converters import validatePythonValue
 from freecad_stub_gen.generators.common.return_type_converter.arg_types import InvalidReturnType, \
     UnionArguments
 from freecad_stub_gen.generators.common.return_type_converter.base import ReturnTypeConverterBase
@@ -15,7 +15,7 @@ from freecad_stub_gen.generators.common.return_type_converter.inner_type_list im
 from freecad_stub_gen.generators.common.return_type_converter.inner_type_tuple import \
     ReturnTypeInnerTuple
 from freecad_stub_gen.generators.exceptions.container import exceptionContainer
-from freecad_stub_gen.util import OrderedStrSet
+from freecad_stub_gen.ordered_set import OrderedStrSet
 
 logger = logging.getLogger(__name__)
