+++ lib/freecad_stub_gen/generators/common/return_type_converter/inner_type_dict.py
@@ -2,6 +2,7 @@ import re
 from functools import wraps
 from typing import ParamSpec, Callable, TypeVar
 
+from freecad_stub_gen.cpp_code.converters import removeQuote
 from freecad_stub_gen.generators.common.cpp_function import generateExpressionUntilChar
 from freecad_stub_gen.generators.common.return_type_converter.arg_types import DictArgument, \
     TypedDictGen, ArgumentsIter, RetType
@@ -85,11 +86,11 @@ class ReturnTypeInnerDict(ReturnTypeConverterBase):
             value = self.getExpressionType(funArgs[1], endPos)
             key = funArgs[0]
             if key.startswith('"') and key.endswith('"'):
-                key = key.removeprefix('"').removesuffix('"')
-                tdg.add(key, value)
+                tKey = removeQuote(key)
+                tdg.add(tKey, value)
             else:
-                key = self.getExpressionType(key, endPos)
-                da.add(key, value)
+                tKey = self.getExpressionType(key, endPos)
+                da.add(tKey, value)
 
         if tdg:
             assert not da, "Values in `TypedDict` are mixed with `dict`"
