+++ lib/freecad_stub_gen/generators/exceptions/gen.py
@@ -2,7 +2,7 @@ import more_itertools
 
 from freecad_stub_gen.generators.common.gen_python_api import PythonApiGenerator
 from freecad_stub_gen.generators.exceptions.container import exceptionContainer
-from freecad_stub_gen.module_container import Module
+from freecad_stub_gen.python_code.module_container import Module
 
 
 class ExceptionGenerator(PythonApiGenerator):
