+++ lib/freecad_stub_gen/generators/from_xml/dynamic_property.py
@@ -5,7 +5,7 @@ from typing import Literal
 
 from freecad_stub_gen.generators.common.gen_property.gen_dynamic import DynamicPropertyGenerator
 from freecad_stub_gen.generators.from_xml.base import BaseXmlGenerator
-from freecad_stub_gen.util import readContent
+from freecad_stub_gen.file_functions import readContent
 
 logger = logging.getLogger(__name__)
