+++ lib/freecad_stub_gen/generators/from_xml/method.py
@@ -14,7 +14,7 @@ from freecad_stub_gen.generators.common.gen_method import MethodGenerator
 from freecad_stub_gen.generators.common.names import getClassNameFromNode
 from freecad_stub_gen.generators.common.signature_merger import SignatureMerger
 from freecad_stub_gen.generators.from_xml.base import BaseXmlGenerator
-from freecad_stub_gen.util import toBool
+from freecad_stub_gen.cpp_code.converters import toBool
 
 logger = logging.getLogger(__name__)
