+++ lib/freecad_stub_gen/python_code/__init__.py
@@ -1,0 +1,5 @@
+import textwrap
+
+
+def indent(block, distance=1, indentSize=4):
+    return textwrap.indent(block, ' ' * distance * indentSize)
